VibeBOT_discord/requirements.txt
discord.py
chatterbot==1.0.4
chatterbot-corpus
pytz